fix(vis+infer): align make_ds defaults with trainer; rebuild vis
- infer_pipeline.make_ds defaults now mirror trainer's ds_kw exactly
  (max_offset_m=0.20, max_rot_deg=0.5, max_crop_px=512). When configs
  omit perturbation hparams, vis was using 3× larger jitter than
  training, blowing up the apparent post-correction error.
- train_ps_v3_ddp.use_convnext default → True.
- model_depth.FrustumLocalEncoder uses the global RNG in eval too, so
  offline inference matches the trainer's never-eval-mode val pass.
- ba_multicam_corr.infer_tiles drops the fixed-seed RNG for the same
  reason; demo_app forces fp16 autocast (matches trainer on V100).
- Replace deleted midtrain_vis with scripts/util/vis.py — single
  forward via DataLoader+collate_full+batch (the val-pass code path),
  σ ellipse always drawn, GT→dist→pred arrow chain.

+++ scripts/serving/demo_app.py
@@ -207,7 +207,9 @@ def _run_model(pkt: dict) -> np.ndarray:
     else:
         dist_in = dist[..., :3]
     pad = torch.zeros(1, dist.shape[1], dtype=torch.bool, device=DEVICE)
-    with torch.no_grad(), torch.autocast(device_type='cuda', dtype=torch.bfloat16):
+    # fp16 to match training (--mixed_precision=fp16). bf16 silently runs
+    # as fp32-emulated on V100 (sm_70) and shifts the input distribution.
+    with torch.no_grad(), torch.autocast(device_type='cuda', dtype=torch.float16):
         out = MODEL(img, dist_in, key_padding_mask=pad,
                     vfp=pkt['vfp'].view(1).to(DEVICE),
                     bucket_uvd=pkt['bucket_uvd'].unsqueeze(0).to(DEVICE),
